requirements.txt
Django==1.3.1
South==0.7.3
django-indexer==0.3.0
django-paging==0.2.4
django-sentry==1.13.5
django-templatetag-sugar==0.1
eventlet==0.9.16
greenlet==0.3.2
lockfile==0.9.1
psycopg2==2.4.4
python-daemon==1.6
raven==0.7.0
simplejson==2.3.2
wsgiref==0.1.2
